Allow setting processing workflow type
- Allow setting processing workflow type from watcher configuration.
- Add watcher configuration defaults and validation.
- Remove AutoApproveAIP option from workflow request and pass type.
- Consider workflow type in processing workflow.
- Remove `unspecified` value and update workflow type enum to a string.
- Update DB schema, generate Ent code and add migration.
- Regenerate API and dashboard client code.

+++ internal/persistence/ent/client/client_test.go
@@ -53,7 +53,7 @@ func createWorkflow(
 ) (*db.Workflow, error) {
 	return entc.Workflow.Create().
 		SetTemporalID("12345").
-		SetType(int8(enums.WorkflowTypeCreateAip)).
+		SetType(enums.WorkflowTypeCreateAip).
 		SetStatus(int8(status)). // #nosec G115 -- constrained value.
 		SetSipID(sipID).
 		Save(context.Background())
